requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
